v0.9.6 - Config Menu for: WiFi ON/OFF, SampleUpload ON/OFF, BLE ON/OFF, Saving Battery Mode Reduced/Lowest - Writting in EEPROM the Config Variables values to make them permanent - Bug fixing:   + Button1 presses in firstBoot went to hibernate regardless how long it was pressed   + Going to the menu screen if Button1 is pressed during bootup phase - change error_setup values to be mask-oriented and re-code the firstSetup for OR-error_setup and   warmingUp() to detect dead errors and take actions

+++ src/httpClient.h
@@ -8,4 +8,4 @@ extern RTC_DATA_ATTR int errorsSampleUpts;
 
 String IpAddress2String(const IPAddress& ipAddress);
 uint8_t sendHttpRequest (boolean logsOn, IPAddress server, uint16_t port, String httpRequest);
-uint8_t sendAsyncHttpRequest (boolean logsOn, boolean fromSetup, uint8_t error_setup, IPAddress server, uint16_t port, String httpRequest, uint64_t* whileLoopTimeLeft);
+uint32_t sendAsyncHttpRequest (boolean logsOn, boolean fromSetup, uint32_t error_setup, IPAddress server, uint16_t port, String httpRequest, uint64_t* whileLoopTimeLeft);
